pyproject.toml
[tool.poetry]
name = "snowflakeapi"
version = "2.0"
description = "SnowFlake API Wrapper for Python"
authors = ["DevSynth <synth@snowflakedev.org>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
pytest = "^6.2.4"
pytest-asyncio = "^0.15.1"
aiohttp = "^3.7.4"
python-dotenv = "^0.18.0"
Sphinx = "^4.1.1"

[tool.poetry.dev-dependencies]
black = "^21.7b0"
isort = "^5.9.2"
furo = "^2021.7.5-beta.38"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
